DEV: move wasm work to web worker
CSP is bypassed in web workers, attempt to perform all wasm work in worker

Note, we may still need some CSP changes to load the static JS file from CDNs
that is a bit more workable.
